ndb - prepare code for more than 255 partitions
use MAX_NDB_PARTITIONS.
add error checks of #partitions.
ensure 16bit are used for fragment-ids in places
there only 8 bits are used.

+++ storage/ndb/src/kernel/vm/SimulatedBlock.hpp
@@ -1365,7 +1365,7 @@ struct Hash2FragmentMap
   STATIC_CONST( MAX_MAP = NDB_DEFAULT_HASHMAP_BUCKETS );
   Uint32 m_cnt;
   Uint32 m_fragments;
-  Uint8 m_map[MAX_MAP];
+  Uint16 m_map[MAX_MAP];
   Uint32 nextPool;
   Uint32 m_object_id;
 };
